Add edit test
